# wegalaxy-foundation/docs/swagger.yaml
basePath: /v1/wegalaxy-foundation
definitions:
...
  model.GetUserInfoRsp:
    properties:
      data:
        $ref: '#/definitions/model.User'
      errcode:
        type: integer
      errmsg:
        type: string
    type: object
...
  model.User:
    properties:
      did:
        type: string
      id:
        type: string
      roles:
        items:
          $ref: '#/definitions/model.UserRole'
        type: array
      status:
        type: integer
      tags:
        items:
          $ref: '#/definitions/model.UserTag'
        type: array
      user_id:
        type: string
    type: object
  model.UserRole:
    properties:
      id:
        type: string
      image:
        type: string
      name:
        type: string
      status:
        type: integer
    type: object
  model.UserTag:
    properties:
      id:
        type: string
      name:
        type: string
      status:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: WeGalaxy Foundation Service
  title: WeGalaxy Foundation Service
  version: "1.0"
paths:
...
  /users:
    get:
      description: Get user information by ID
      operationId: get-user
      parameters:
      - description: Token passed in Authorization in 'Bearer xxxx' format. The alternative
          is to pass 'token=xxxx' in cookie
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetUserInfoRsp'
      summary: Get user information
      tags:
      - user
...
swagger: "2.0"